completed admin endpoint

+++ src/controllers/admin/projectController.js
@@ -5,7 +5,7 @@ exports.getAll = (req, res) => {
     db.query("SELECT * FROM `Internal_project`", (err, result) => {
         if (err) return res.status(400).json({
             success: false,
-            error_message: "admin get all internal projects failed"
+            error_message: "admin get all internal projects failed; database error"
         });
 
         const projects = JSON.parse(JSON.stringify(result));
@@ -24,15 +24,31 @@ exports.modify = (req, res) => {
     Object.keys(modified).forEach(key => modified[key] === undefined && delete modified[key]);
 
     verifyAdmin(admin_key, () => {
-        db.query("UPDATE `Internal_project` SET ? WHERE idx = ?", [modified, project_idx], (err, result) => {
+        // Check whether the project is currently ongoing
+        db.query("SELECT * FROM `Current_project` WHERE project_idx = ?", project_idx, (err, result) => {
             if (err) return res.status(400).json({
                 success: false,
-                error_message: "admin modify project failed"
+                error_message: "admin modify team failed; database error"
             });
 
-            res.status(200).json({
-                success: true
-            });
+            const current_projects = JSON.parse(JSON.stringify(result));
+            if (current_projects.length !== 0) {
+                res.status(400).json({
+                    success: false,
+                    error_message: "admin modify project failed; project currently ongoing"
+                });
+            } else {
+                db.query("UPDATE `Internal_project` SET ? WHERE idx = ?", [modified, project_idx], (err, result) => {
+                    if (err) return res.status(400).json({
+                        success: false,
+                        error_message: "admin modify project failed; database error"
+                    });
+
+                    res.status(200).json({
+                        success: true
+                    });
+                });
+            }
         });
     });
 };
@@ -41,15 +57,31 @@ exports.delete = (req, res) => {
     const {admin_key, project_idx} = req.body;
 
     verifyAdmin(admin_key, () => {
-        db.query("DELETE FROM `Internal_project` WHERE idx = ?", project_idx, (err, result) => {
+        // Check whether the project is currently ongoing
+        db.query("SELECT * FROM `Current_project` WHERE project_idx = ?", project_idx, (err, result) => {
             if (err) return res.status(400).json({
                 success: false,
-                error_message: "admin delete project failed"
+                error_message: "admin modify team failed; database error"
             });
 
-            res.status(200).json({
-                success: true
-            });
+            const current_projects = JSON.parse(JSON.stringify(result));
+            if (current_projects.length !== 0) {
+                res.status(400).json({
+                    success: false,
+                    error_message: "admin modify project failed; project currently ongoing"
+                });
+            } else {
+                db.query("DELETE FROM `Internal_project` WHERE idx = ?", project_idx, (err, result) => {
+                    if (err) return res.status(400).json({
+                        success: false,
+                        error_message: "admin delete project failed; database error"
+                    });
+
+                    res.status(200).json({
+                        success: true
+                    });
+                });
+            }
         });
     });
 };
